// auctionhouseapp.client/src/dto/ILoginUserInfo.d.ts
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface ILoginUserInfo
{
	loginUserId: string;
	loginUserName: string;
	userType: string;
	roleList: string[];
	status: string;
	expiresTime: string;
	emailAddr: string;
	phone: string;
}